backend/requirements.txt
django>=4.2,<6.0
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0
